+++ src/index.ts
@@ -3,5 +3,5 @@
  * via shared interface.
  * @module git-smelt
  */
-export { github } from './github';
-export { gitlab } from './gitlab';
+export * from './github';
+export * from './gitlab';